+++ back-to-monitor@nathan818fr/.prettierrc.yaml
@@ -1,0 +1,9 @@
+trailingComma: es5
+printWidth: 120
+tabWidth: 4
+semi: true
+singleQuote: true
+bracketSpacing: false
+overrides:
+    - files: ['*.json']
+      options: {tabWidth: 2}
