--- back-to-monitor@nathan818fr/files/back-to-monitor@nathan818fr/5.4/settings-schema.json
@@ -1,10 +1,0 @@
-{
-  "incompatibilityNotice1": {
-    "type": "header",
-    "description": "This extension is not compatible with this version of Cinnamon."
-  },
-  "incompatibilityNotice2": {
-    "type": "header",
-    "description": "Cinnamon now natively provides the features of this extension."
-  }
-}
